+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Yii Framework 2 Change Log
 
 1.4.0
 -----
+- Enh: backend user view page enhancement
+- Enh #29: delete button on log record page
 - Enh #28: init tool + local config files
 - Fix #25: backend user update
 - Chg: added utf-8 charset to nginx.conf